refactor(UserAuth) : UserAuth 모델을 dao로 활용하도록 변경
password 비교 메서드 추가
UserSession 파싱 메서드 추가
